Testing Q2=2.11,3.0,5.5

+++ src/models/lt_2D_fit.py
@@ -3,7 +3,7 @@
 #
 # Description:
 # ================================================================
-# Time-stamp: "2024-07-19 15:13:24 trottar"
+# Time-stamp: "2024-07-19 15:14:16 trottar"
 # ================================================================
 #
 # Author:  Richard L. Trotta III <trotta@cua.edu>
@@ -16,6 +16,7 @@ from ROOT import TGraphErrors, TF1, TF2, TGraph2DErrors, TCanvas
 from ROOT import TString, TNtuple, TMinuit
 from array import array
 import math
+import ctypes
 import os, sys
 
 ParticleType = sys.argv[1]
@@ -211,7 +212,7 @@ def single_setting(q2_set, w_set, fn_lo, fn_hi):
 
         for ii in range(glo.GetN()):
             
-            glo.GetPoint(ii, g_xx, g_yy)
+            glo.GetPoint(ii, ctypes.c_double(g_xx), ctypes.c_double(g_yy)
             g_yy_err = math.sqrt((glo.GetErrorY(ii) / g_yy)**2 + (pt_to_pt_systematic_error/100)**2) * g_yy
 
             lo_cross_sec_err[i] += 1 / (g_yy_err**2)
@@ -221,7 +222,7 @@ def single_setting(q2_set, w_set, fn_lo, fn_hi):
 
         for ii in range(ghi.GetN()):
             
-            ghi.GetPoint(ii, g_xx, g_yy)
+            ghi.GetPoint(ii, ctypes.c_double(g_xx), ctypes.c_double(g_yy)
             g_yy_err = math.sqrt((ghi.GetErrorY(ii) / g_yy)**2 + (pt_to_pt_systematic_error/100)**2) * g_yy
 
             hi_cross_sec_err[i] += 1 / (g_yy_err**2)
